src/main/webapp/app/entities/catalog/list/catalog.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpHeaders, HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';

import { CatalogService } from '../service/catalog.service';

import { CatalogComponent } from './catalog.component';

describe('Component Tests', () => {
  describe('Catalog Management Component', () => {
    let comp: CatalogComponent;
    let fixture: ComponentFixture<CatalogComponent>;
    let service: CatalogService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [CatalogComponent],
      })
        .overrideTemplate(CatalogComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(CatalogComponent);
      comp = fixture.componentInstance;
      service = TestBed.inject(CatalogService);

      const headers = new HttpHeaders().append('link', 'link;link');
      spyOn(service, 'query').and.returnValue(
        of(
          new HttpResponse({
            body: [{ idCatalog: 123 }],
            headers,
          })
        )
      );
    });

    it('Should call load all on init', () => {
      // WHEN
      comp.ngOnInit();

      // THEN
      expect(service.query).toHaveBeenCalled();
      expect(comp.catalogs[0]).toEqual(jasmine.objectContaining({ idCatalog: 123 }));
    });

    it('should load a page', () => {
      // WHEN
      comp.loadPage(1);

      // THEN
      expect(service.query).toHaveBeenCalled();
      expect(comp.catalogs[0]).toEqual(jasmine.objectContaining({ idCatalog: 123 }));
    });

    it('should calculate the sort attribute for an id', () => {
      // WHEN
      comp.ngOnInit();

      // THEN
      expect(service.query).toHaveBeenCalledWith(expect.objectContaining({ sort: ['idCatalog,asc'] }));
    });

    it('should calculate the sort attribute for a non-id attribute', () => {
      // INIT
      comp.ngOnInit();

      // GIVEN
      comp.predicate = 'name';

      // WHEN
      comp.loadPage(1);

      // THEN
      expect(service.query).toHaveBeenLastCalledWith(expect.objectContaining({ sort: ['name,asc', 'idCatalog'] }));
    });

    it('should re-initialize the page', () => {
      // WHEN
      comp.loadPage(1);
      comp.reset();

      // THEN
      expect(comp.page).toEqual(0);
      expect(service.query).toHaveBeenCalledTimes(2);
      expect(comp.catalogs[0]).toEqual(jasmine.objectContaining({ idCatalog: 123 }));
    });
  });
});
